pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Miyamura80", email = "eitomiyamura@gmail.com" }
]
dependencies = [
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.1",
    "human-id>=0.2.0",
    "import-linter>=2.1.1",
    "pytest>=8.3.4",
    "pytest-xdist>=3.6.1",
    "termcolor>=2.5.0",
    "loguru>=0.7.3",
    "vulture>=2.14",
    "dspy>=3.1.2",
    "langfuse>=3.12.1",
    "litellm>=1.59.8",
    "tenacity>=9.0.0",
    "pillow>=12.1.0",
    "google-genai>=1.60.0",
    "ty>=0.0.14",
    "pytest-env>=1.2.0",
    "pydantic-settings>=2.7.1",
    "pytest-cov>=7.0.0",
    "pytest-repeat>=0.9.3",
    "pylint>=3.3.3",
    "deptry>=0.23.0",
    "openfeature-sdk>=0.7.1",
    "scrubadub>=2.0.1",
    "pydantic>=2.10.6",
    "numpy>=2.2.2",
    "typer",
    "questionary",
    "rich>=14.3.1",
    # SaaS-specific dependencies
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.42",
    "requests>=2.32.4",
    "pytest-asyncio>=1.2.0",
    "fastapi>=0.118.0",
    "PyJWT>=2.8.0",
    "uvicorn>=0.37.0",
    "itsdangerous>=2.2.0",
    "stripe>=13.0.1",
    "workos>=4.0.0",
    "httpx>=0.27.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src", "common", "utils"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "C901", "FIX"]
ignore = ["E501", "UP015", "B008"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ty]

[tool.ty.rules]
possibly-unresolved-reference = "error"

[tool.ty.terminal]
error-on-warning = true

[tool.ty.environment]
python-version = "3.12"

[tool.deptry]
ignore = ["DEP002"]

[tool.vulture]
exclude = [
    ".venv/",
    ".uv-cache/",
    ".uv_cache/",
    ".uv_tools/",
    ".uv-tools/",
    "docs/",
    "node_modules/",
    "tests/test_logging_security.py",
    "tests/test_logging_thread_safety.py",
    "tests/test_template.py",
    "tests/e2e/e2e_test_base.py",
    "tests/e2e/payments/test_stripe.py",
    "tests/e2e/agent/tools/test_alert_admin.py",
    "utils/llm/",
    "common/",
    "src/utils/logging_config.py",
    "src/utils/context.py",
    "tests/conftest.py",
    "init/",
    "onboard.py",
    "alembic/",
    "src/db/",
    "src/api/routes/",
    "src/api/auth/",
    "src/utils/integration/",
    "src/stripe/",
    "scripts/"
]
ignore_names = [
    "apply_referral",
    "referral_count",
    "get_or_create_referral_code"
]

[tool.coverage.run]
branch = true
source = ["src", "common", "utils"]
omit = [
    "tests/*",
    "docs/*",
    "init/*",
    "alembic/*",
]

[tool.coverage.report]
show_missing = true

[dependency-groups]
dev = [
    "pytest-check-links>=0.9.1",
]
